+++ CHANGELOG.md
@@ -34,6 +34,8 @@
 * [CHANGE] Remove middleware package. #182
 * [CHANGE] Memberlist: disabled TCP-based ping fallback, because dskit already uses a custom transport based on TCP. #194
 * [CHANGE] Memberlist: KV store now fast-joins memberlist cluster before serving any KV requests. #195
+* [CHANGE] Add inet6 support for ring util. #185
+* [CHANGE] Memberlist: Add inet6 preference support. #185
 * [ENHANCEMENT] Add middleware package. #38
 * [ENHANCEMENT] Add the ring package #45
 * [ENHANCEMENT] Add limiter package. #41